refactor: address review feedback from leongdl
- Simplify _matches_any_filter() to use any() expression
- Move _validate_and_normalize_include_paths() to _job_download_helpers.py
- Skip TTY stdin reopen in JSON mode since there are no interactive prompts
